Fix migration
Former-commit-id: 5a8cd3eed71b3a8ba28441ebb63e9f7a5bcfc26a

+++ db/migrate/20211214095858_remove_initials_from_projects.rb
@@ -1,6 +1,9 @@
 class RemoveInitialsFromProjects < ActiveRecord::Migration[5.2]
   def change
-    remove_column :projects, :initials, :string
+    if ActiveRecord::Base.connection.column_exists?(:projects, :initials)
+      remove_column :projects, :initials, :string
+    end
+
     Project.all.each do |project|
       project.update_attributes shortname: project.shortname.downcase
     end
